Improve layout for ComparisonTable, update copy, fix tests (#556)
* Add missing key prop

* Remove copy from CompanyView to match latest suggestion

* Update table headers and add some strings to translations

* Improve table header font size for small screen sizes

* Improve size and spacing for table

* Use full-width table for all mobile sizes.

* Fix offset for sticky table header. Also hide background content above it.

* Improve compact layout for tables on smaller screens

* Improve table layout for small screens.

* More compact table header to save screen height on mobile

* Cover full width with table for all mobile devices

* Fix margin positioning for ToggleButton

* Update table footer copy

* Reference arrow svg as img instead of inlined SVG. This improves rendering performance.

* Update newsletter info

* Update about us text

* Move text about previous project

* Update sources and method

* Use plural key to prepare for i18n

* Fix tests for StartPage by splitting into two modules: CompanyView and RegionalView

* Cleanup tests

* Add privacy policy as separate page

* Add description of the method for company data
