+++ components/PageWrapper.tsx
@@ -17,11 +17,11 @@ const WrapInner = styled.div<{ compact?: boolean }>`
   ${({ compact }) => (compact ? css`
     padding: 0;
 
-    @media only screen and (${devices.smallMobile}) {
-      padding: 0px 16px;
+    @media only screen and (${devices.tablet}) {
+      padding: 0 16px;
     }
   ` : css`
-    padding: 0px 16px;
+    padding: 0 16px;
   `)}
 
   @media only screen and (${devices.tablet}) {
