+++ components/ToggleButton.tsx
@@ -2,8 +2,9 @@ import styled from 'styled-components'
 import { devices } from '../utils/devices'
 
 const ToggleBtn = styled.button`
-  height: 32px;
+  height: 36px;
   margin: 16px;
+  margin-right: 4px;
   position: absolute;
   top: 0;
   right: 0;
@@ -17,8 +18,8 @@ const ToggleBtn = styled.button`
   justify-content: right;
   z-index: 150;
 
-  @media only screen and (${devices.mobile}) {
-    margin: 12px;
+  @media only screen and (${devices.smallMobile}) {
+    margin-right: 12px;
   }
 `
