+++ public/locales/sv/about.json
@@ -38,11 +38,11 @@
     "title": "Vår styrelse"
   },
   "content": {
-    "text": "I dag kan du på Klimatkollen se:\n\n- Koldioxidbudgetar för landets alla 290 kommuner. Klimatkollen utgår ifrån en [nationell koldioxidbudget](https://www.cemus.uu.se/wp-content/uploads/2023/12/Paris-compliant-carbon-budgets-for-Swedens-counties-.pdf) som beräknats av forskare vid Uppsala universitet enligt Tyndall-modellen och som sedan fördelats ut på kommunerna av Klimatkollen. (2022 koldioxidbudget [här](/Paris_compliant_Swedish_CO2_budgets-March_2022-Stoddard&Anderson.pdf))\n- En översikt i både kartvy och listvy över koldioxidutsläppen i landets kommuner. Kommunerna rankas baserat på genomsnittlig årlig utsläppsminskningstakt sedan Parisavtalet undertecknades 2015.\n- En visualisering av hur det gått med koldioxidutsläppen för varje kommun från 1990 tills idag, hur det borde gå för att klara Parisavtalets 1,5-gradersmål och en prognos för hur det går om utsläppen i kommunen fortsätter att utvecklas som de senaste fem åren.\n- För varje kommun finns även annan nyckelfakta, som koldioxidbudget, datum när budgeten tar slut med nuvarande utsläppstakt, politiskt styre och koldioxidutsläpp per invånare.\n\nKlimatkollen är utvecklad med öppen källkod. Det betyder att alla kan vara med och utveckla och förbättra sajten via vårt [Github-repo](https://github.com/Klimatbyran/klimatkollen).",
+    "text": "I dag kan du på Klimatkollen se:\n\n- Koldioxidbudgetar för landets alla 290 kommuner. Klimatkollen utgår ifrån en [nationell koldioxidbudget](https://www.cemus.uu.se/wp-content/uploads/2023/12/Paris-compliant-carbon-budgets-for-Swedens-counties-.pdf) som beräknats av forskare vid Uppsala universitet enligt Tyndall-modellen och som sedan fördelats ut på kommunerna av Klimatkollen. (2022 koldioxidbudget [här](/Paris_compliant_Swedish_CO2_budgets-March_2022-Stoddard&Anderson.pdf))\n- En översikt i både kartvy och listvy över koldioxidutsläppen i landets kommuner. Kommunerna rankas baserat på genomsnittlig årlig utsläppsminskningstakt sedan Parisavtalet undertecknades 2015.\n- En visualisering av hur det gått med koldioxidutsläppen för varje kommun från 1990 tills idag, hur det borde gå för att klara Parisavtalets 1,5-gradersmål och en prognos för hur det går om utsläppen i kommunen fortsätter att utvecklas som de senaste fem åren.\n- För varje kommun finns även annan nyckelfakta, som koldioxidbudget, datum när budgeten tar slut med nuvarande utsläppstakt, politiskt styre och koldioxidutsläpp per invånare.\n- Utsläppsdata om Sveriges 150 största företag för 2023, avgränsade till Stockholmsbörsens storbolagslista samt de 15 största statliga bolagen, de fem största ekonomiska föreningarna samt IKEA och ICA Gruppen.\n\nKlimatkollen är utvecklad med öppen källkod. Det betyder att alla kan vara med och utveckla och förbättra sajten via vårt [Github-repo](https://github.com/Klimatbyran/klimatkollen).",
     "title": "Vad finns på Klimatkollen?"
   },
   "earlierProjects": {
-    "text": "Inför riksdagsvalet 2022 ansvarade Klimatkollen för två unika projekt för att ge medborgare bättre koll på utsläppseffekterna av partiernas klimatpolitik:\n\n1. Klimatmål\nEn [analys](/partierna) av riksdagspartiernas klimatmål, tillsammans med forskarnätverket Researchers’ Desk, Världsnaturfonden WWF, Våra barns klimat och ClimateView, i samarbete med PwC och Naturskyddsföreningen. Analysen visade att sex av åtta partier missar Parisavtalets 1,5-gradersmål.\n2. [Utsläppsberäkning](/utslappsberakningar) av riksdagspartiernas politik gällande tolv centrala klimatåtgärder. Bakom uträkningarna står Klimatkollen, Världsnaturfonden WWF, ClimateView, Naturskyddsföreningen och Våra barns klimat. Beräkningarna visade att den nya regeringens politik kan öka utsläppen med 25 miljoner ton redan under mandatperioden 2022–2026.",
+    "text": "Under 2023 har Klimatkollen uppburit stöd av [Postkodstiftelsen](https://postkodstiftelsen.se/) för att ta tempen på hur det går med klimatomställningen i kommunerna. I dag kan du se hur det går med koldioxidutsläppen. Nu visar vi även vad kommunerna gör för att minska dem. Vi utgår ifrån de största utsläppssektorerna och visar sedan hur det går med viktiga omställningar inom dessa sektorer, genom att presentera statistik om utvalda nyckeltal, ett nyckeltal per omställning.\n\nInför riksdagsvalet 2022 ansvarade Klimatkollen för två unika projekt för att ge medborgare bättre koll på utsläppseffekterna av partiernas klimatpolitik:\n\n1. Klimatmål\nEn [analys](/partierna) av riksdagspartiernas klimatmål, tillsammans med forskarnätverket Researchers’ Desk, Världsnaturfonden WWF, Våra barns klimat och ClimateView, i samarbete med PwC och Naturskyddsföreningen. Analysen visade att sex av åtta partier missar Parisavtalets 1,5-gradersmål.\n2. [Utsläppsberäkning](/utslappsberakningar) av riksdagspartiernas politik gällande tolv centrala klimatåtgärder. Bakom uträkningarna står Klimatkollen, Världsnaturfonden WWF, ClimateView, Naturskyddsföreningen och Våra barns klimat. Beräkningarna visade att den nya regeringens politik kan öka utsläppen med 25 miljoner ton redan under mandatperioden 2022–2026.",
     "title": "Tidigare projekt"
   },
   "funding": {
@@ -54,10 +54,6 @@
     "description": "Om Klimatkollen"
   },
   "policies": "Här hittar du våra [stadgar](/stadgar.pdf), [uppförandekod](/uppforandekod.pdf) och [antikorruptionspolicy](/antikorruptionspolicy.pdf).",
-  "postcode": {
-    "text": "Under 2023 har Klimatkollen uppburit stöd av [Postkodstiftelsen](https://postkodstiftelsen.se/) för att ta tempen på hur det går med klimatomställningen i kommunerna. I dag kan du se hur det går med koldioxidutsläppen. Nu visar vi även vad kommunerna gör för att minska dem. Vi utgår ifrån de största utsläppssektorerna och visar sedan hur det går med viktiga omställningar inom dessa sektorer, genom att presentera statistik om utvalda nyckeltal, ett nyckeltal per omställning.",
-    "title": "Postkodprojektet"
-  },
   "problem": {
     "text": "Det finns mycket data därute som kan hjälpa oss att ta tempen på klimatomställningen, men den är ofta komplex, svårförståelig eller inlåst bakom betalvägg. Myndigheternas statistik om klimatutsläppen i kommunerna släpar efter med nästan två års fördröjning. Så kan vi inte ha det!\n\nVi jobbar för att göra klimatdata lättillgänglig för medborgare. Det möjliggör ansvarsutkrävande och ger bränsle till nya digitala innovationer som snabbar på klimatomställningen. Vi behöver veta mer om utsläppen, varifrån de kommer, vilka lösningar som genomförs i vilka kommuner och vilken effekt dessa åtgärder kan få. Läs mer om det i [vår debattartikel](https://www.aktuellhallbarhet.se/alla-nyheter/debatt/med-battre-klimatdata-kan-vi-accelerera-klimatomstallningen/).",
     "title": "Problemet vi löser"
