+++ public/locales/sv/kallorMetod.json
@@ -16,7 +16,7 @@
     "title": "Klimatkollen utgår från Parisavtalet"
   },
   "sources": {
-    "text": "Klimatkollen baseras på offentliga källor och verifierad data. Vi anger alla källor så att du enkelt kan kolla upp och läsa mer. Om något blivit fel, mejla oss gärna på [hej@klimatkollen.se](mailto:hej@klimatkollen.se) så att vi kan ändra.\n\nStatistik om koldioxidutsläpp hämtas från [Nationella emissionsdatabasen](https://nationellaemissionsdatabasen.smhi.se/), som hanteras av SMHI och utgår från Sveriges officiella utsläppsstatistik, med Naturvårdsverket som ansvarig myndighet. Kommunal utsläppsdata har 1,5–2 års fördröjning, 2021 års siffror släpptes i juli 2023.\n\nUppgifter om politiskt styre i kommunerna hämtas från [Sveriges Kommuner och Regioner](https://skr.se/skr/demokratiledningstyrning/valmaktfordelning/valresultatstyren/styrekommunereftervalet2022.69547.html). Från [Wikidata](https://www.wikidata.org/wiki/Wikidata:Country_subdivision_task_force/Sweden/Municipalities) hämtar vi kommunernas invånarantal och kommunvapen. Statistiken om andel laddbara bilar kommer från [Trafikanalys](https://www.trafa.se/vagtrafik/fordon/). Cykelvägsdata tas från Trafikverkets [Nationella vägdatabas](https://nvdb2012.trafikverket.se/SeTransportnatverket). För hushållens konsumtionsbaserade utsläpp kommer data från [Stockholm Environment Institute](https://www.sei.org/tools/konsumtionskompassen/). Uppgifter om laddinfrastrukturen kommer från branschorganisationen [Power Circles databas ELIS](https://powercircle.org/elbilsstatistik/). Kommunernas [klimatplaner](https://docs.google.com/spreadsheets/d/13CMqmfdd6QUD6agKFyVhwZUol4PKzvy253_EwtsFyvw/edit#gid=0) har vi samlat in med hjälp av allmänheten och Klimatklubbens och Naturskyddsföreningens medlemmar. Information om klimatkrav vid kommunernas upphandlingar kommer från [Upphandlingsmyndigheten](https://www.klimatkollen.se/data/procurements/NUE2022_DATA_2023-12-20.xlsx) och [Greenpeace](https://docs.google.com/spreadsheets/d/1EdHUa49HJZn0rXqM-6tChdim4TJzXnwA/edit#gid=1040317160).",
+    "text": "Klimatkollen baseras på offentliga källor och verifierad data. Vi anger alla källor så att du enkelt kan kolla upp och läsa mer. Om något blivit fel, mejla oss gärna på [hej@klimatkollen.se](mailto:hej@klimatkollen.se) så att vi kan ändra.\n\nFöretagsdata hämtas från bolagens egna årsredovisningar och hållbarhetsrapporter. För metodbeskrivning läs [här](/metodbeskrivning-foretagsdata.pdf)\n\nStatistik om kommunernas koldioxidutsläpp hämtas från [Nationella emissionsdatabasen](https://nationellaemissionsdatabasen.smhi.se/), som hanteras av SMHI och utgår från Sveriges officiella utsläppsstatistik, med Naturvårdsverket som ansvarig myndighet. Kommunal utsläppsdata har 1,5–2 års fördröjning, 2021 års siffror släpptes i juli 2023.\n\nUppgifter om politiskt styre i kommunerna hämtas från [Sveriges Kommuner och Regioner](https://skr.se/skr/demokratiledningstyrning/valmaktfordelning/valresultatstyren/styrekommunereftervalet2022.69547.html). Från [Wikidata](https://www.wikidata.org/wiki/Wikidata:Country_subdivision_task_force/Sweden/Municipalities) hämtar vi kommunernas invånarantal och kommunvapen. Statistiken om andel laddbara bilar kommer från [Trafikanalys](https://www.trafa.se/vagtrafik/fordon/). Cykelvägsdata tas från Trafikverkets [Nationella vägdatabas](https://nvdb2012.trafikverket.se/SeTransportnatverket). För hushållens konsumtionsbaserade utsläpp kommer data från [Stockholm Environment Institute](https://www.sei.org/tools/konsumtionskompassen/). Uppgifter om laddinfrastrukturen kommer från branschorganisationen [Power Circles databas ELIS](https://powercircle.org/elbilsstatistik/). Kommunernas [klimatplaner](https://docs.google.com/spreadsheets/d/13CMqmfdd6QUD6agKFyVhwZUol4PKzvy253_EwtsFyvw/edit#gid=0) har vi samlat in med hjälp av allmänheten och Klimatklubbens och Naturskyddsföreningens medlemmar. Information om klimatkrav vid kommunernas upphandlingar kommer från [Upphandlingsmyndigheten](https://www.klimatkollen.se/data/procurements/NUE2022_DATA_2023-12-20.xlsx) och [Greenpeace](https://docs.google.com/spreadsheets/d/1EdHUa49HJZn0rXqM-6tChdim4TJzXnwA/edit#gid=1040317160).",
     "title": "Om våra källor"
   },
   "title": "Källor och metod"
