+++ public/locales/sv/privacy.json
@@ -1,0 +1,3 @@
+{
+  "privacyPolicy": "# Hur Klimatkollen behandlar personuppgifter\n\n*2024-04-19*\n\n## Personuppgiftsansvar\nKlimatbyrån med organisationsnummer 802537-8814 är personuppgiftsansvarig när föreningens verksamheter behandlar personuppgifter.\n\n## Vilka uppgifter behandlar vi?\nDe personuppgifter vi samlar in är i första hand sådana som du själv uppger (namn, adress, mobilnummer etc.) när du exempelvis skriver upp dig för vårt nyhetsbrev eller blir medlem. Du kan också vara anställd hos en samarbetspartner eller annan kontakt till Klimatbyrån, då är personuppgifterna dina kontaktuppgifter i arbetet.\n\n## Varför behandlar vi personuppgifterna?\nVi behandlar uppgifterna för att kunna fullfölja de åtaganden som är kopplade till det syfte som uppges vid insamlingstillfället eller för samarbetet. I övrigt använder vi dina uppgifter endast för ändamål som rör Klimatbyråns egen verksamhet. Vi lämnar inte ut uppgifterna till tredje part.\n\nKlimatbyrån värnar om den personliga integriteten. Att hantera personuppgifter korrekt och säkert är en prioriterad fråga för oss. De som anförtror oss sina uppgifter ska vara trygga.\n\nVi följer Dataskyddsförordningen och de andra lagar och regler som gäller på dataskyddsområdet. Du har alltid rätt att begära korrigering eller radering av dina personliga uppgifter.\n\n## Frågor\nHar du frågor eller önskemål kring hantering av dina personuppgifter, vänligen kontakta oss via <hej@klimatkollen.se>.\n\n## In English\nIf you have questions or requests about your personal data, please contact us at <hej@klimatkollen.se>."
+}
